Porting FreeRTOS files

+++ SYSTEM/delay/delay.h
@@ -3,7 +3,7 @@
 #include "sys.h"  
  
 void delay_init(void);
-void delay_ms(u16 nms);
+void delay_ms(u32 nms);
 void delay_us(u32 nus);
 
 #endif
